refactor: move LetDecl in Decl directory
